Enable import/export into folders
Previously it was not possible to export and import from and to folders.
In order to achieve this a new parameter `--folder` is introduced.
Given this parameter it will resolve the URL of the folder and
export/import jobs there.

In order to achieve this, the API endpoint /api/xml of a folder is used
(also the for the "root" folder) instead of /view/all/api/xml in order
not to rely on special view, which is only existing in root.

Also a bug is fixed if an old version of Jenkins does not
support crumb issueing.