Add credential management (decrypt and apply)
This enables migration of credentials between Jenkins instances.

Butler will get the XML config of any folder and will ingest
the encrypted secrets. The encrypted secrets are then decrypted
via a Groovy script executed on Jenkins master. This needs
most of the times admin privileges.

In a second step you can pipe in the decrypted credentials
and apply them to another folder or even instance.
Also a Groovy script is leveraged in order to achieve this.